% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{author_network}
\alias{author_network}
\title{Author collaboration network as of \code{"2022-07-03 14:37:31 UTC"}}
\format{
An object of class \code{cranly_network} (inherits from \code{list}) of length 2.
}
\usage{
author_network
}
\description{
The author collaboration network as of \code{"2022-07-03 14:37:31 UTC"},
after cleaning and organizing package and author names using
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}
}
\seealso{
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}, \link{cran_db}, \link{package_network}
}
\keyword{datasets}
