% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\name{build_dependence_tree}
\alias{build_dependence_tree}
\title{\code{build_dependence_tree} method for an object}
\usage{
build_dependence_tree(x, ...)
}
\arguments{
\item{x}{an object to use for building a dependence tree}

\item{...}{other arguments to be passed to the method}
}
\description{
\code{build_dependence_tree} method for an object
}
\seealso{
build_network.cranly_network compute_dependence_tree
}
