% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network.cranly_db}
\alias{build_network.cranly_db}
\alias{cranly_network}
\title{Compute edges and nodes of package and author networks}
\usage{
\method{build_network}{cranly_db}(object = clean_CRAN_db(), trace = FALSE,
  perspective = "package", ...)
}
\arguments{
\item{object}{a \code{\link{cranly_db}} object}

\item{trace}{logical. Print progress infromation? Default is \code{FALSE}}

\item{perspective}{character. Build a \code{"package"} (default )or an \code{"author"} network?}

\item{...}{Currently not used}
}
\description{
Compute edges and nodes of package and author networks
}
\examples{
\dontrun{
package_db <- clean_CRAN_db()
package_network <- build_network(package_db)
head(package_network$edges)
head(package_network$nodes)
attr(package_network, "timestamp")
class(package_network)
}

}
