% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\name{build_network}
\alias{build_network}
\title{\code{build_network} method for an object}
\usage{
build_network(object, ...)
}
\arguments{
\item{object}{an object to use for building a network}

\item{...}{other arguments to be passed to the method}
}
\description{
\code{build_network} method for an object
}
\seealso{
build_network.cranly_network
}
