% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_CRAN_db.R
\name{clean_up_author}
\alias{clean_up_author}
\title{Clean up author names}
\usage{
clean_up_author(variable)
}
\arguments{
\item{variable}{a character string.}
}
\value{
A list of one vector of character strings.
}
\description{
Clean up author names
}
\examples{
clean_up_author(paste("The R Core team, Brian & with some assistance from Achim, Hadley;",
                      "Kurt\\n Portugal; Ireland; Italy; Greece; Spain"))
}
