% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran20032018}
\alias{cran20032018}
\title{Cleaned and processed CRAN package metadata from 20 March 2018}
\format{An object of class \code{cranly_db} (inherits from \code{data.frame}) with 12337 rows and 13 columns.}
\usage{
cran20032018
}
\description{
This is a processed version of the result of a
\code{\link[tools]{CRAN_package_db}()} call on 20 March
2018. Specifically, the resulting data frame was passed to the
\code{\link{clean_CRAN_db}} function and only the following
variables have been kept:
\itemize{
  \item package
  \item version
  \item date
  \item url
  \item depends
  \item imports
  \item suggests
  \item enhances
  \item license
  \item author
  \item maintainer
  \item contact
  \item md5sum
}
}
\details{
The author names have been converted to ASCII using transliteration
through the \code{\link{iconv}} function.
}
\keyword{datasets}
