% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_db}
\alias{cran_db}
\title{A dump of a state of the CRAN package database after cleaning on \code{"2022-07-03 14:37:31 UTC"}}
\format{
An object of class \code{cranly_db} (inherits from \code{data.frame}) with 18346 rows and 67 columns.
}
\usage{
cran_db
}
\description{
The CRAN package database as of \code{"2022-07-03 14:37:31 UTC"}, after
cleaning and organizing package and author names using
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}
}
\seealso{
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}, \link{package_network}, \link{author_network}
}
\keyword{datasets}
