% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crandb}
\alias{crandb}
\title{A dump of a state of the CRAN package database after cleaning}
\format{
An object of class \code{cranly_db} (inherits from \code{data.frame}) with 18320 rows and 67 columns.
}
\usage{
crandb
}
\description{
The CRAN package database as of \code{"2022-06-30 11:12:25 BST"}, after
cleaning and organizing package and author names using
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}
}
\seealso{
\code{\link[=clean_CRAN_db]{clean_CRAN_db()}}
}
\keyword{datasets}
