% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\docType{package}
\name{cranly}
\alias{cranly}
\alias{cranly-package}
\title{cranly: CRAN package database analytics and visualizations}
\description{
cranly: CRAN package database analytics and visualizations
}
