% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\name{package_by}
\alias{package_by}
\alias{package_with}
\alias{author_with}
\alias{author_of}
\alias{suggests}
\alias{imports}
\alias{depends}
\alias{linking_to}
\alias{enhances}
\alias{maintainer_of}
\alias{maintained_by}
\title{Find packages, authors, maintainers, license, versions etc by authors, packages or names matching a specific string}
\usage{
package_by(x, author = NULL, exact = FALSE)

package_with(x, name = NULL, exact = FALSE)

author_with(x, name = NULL, exact = FALSE)

author_of(x, package = NULL, exact = FALSE)

suggests(x, package = NULL, exact = FALSE)

imports(x, package = NULL, exact = FALSE)

depends(x, package = NULL, exact = FALSE)

linking_to(x, package = NULL, exact = FALSE)

enhances(x, package = NULL, exact = FALSE)

maintainer_of(x, package = NULL, exact = FALSE)

maintained_by(x, name = NULL, exact = FALSE)
}
\arguments{
\item{x}{a \code{\link{cranly_network}} object}

\item{author}{a vector of character strings with the author names to be matched. Default is \code{Inf} which returns all available author in \code{x} for further subsetting}

\item{exact}{logical. Should we use exact matching? Default is \code{TRUE}}

\item{name}{a vector of character strings with the names to be matched. If \code{Inf} all available names in \code{x} are returned. If \code{NULL} (default) nothing is matched}

\item{package}{a vector of character strings with the package names to be matched. Default is \code{Inf} which returns all available packages in \code{x} for further subsetting}

\item{flat}{if \code{TRUE} (default) then the results is an unnamed character vector. See Details for what each function returns when \code{flat = FALSE}.}
}
\description{
Find packages, authors, maintainers, license, versions etc by authors, packages or names matching a specific string
}
\details{

}
\examples{
\dontrun{
cran_db <- clean_CRAN_db()
## Using a package directives network
package_network <- build_network(cran_db)
## Find all packages containing glm in their name
package_with(package_network, name = "glm")
## Find all authors of packages containing brglm in their name
author_of(package_network, package = "rglm", exact = FALSE)
## Find all packages with brglm in their name
package_with(package_network, name = "rglm", exact = FALSE)
## Find all authors of the package brglm2
author_of(package_network, package = "brglm2", exact = TRUE)
## Find all authors with Ioannis in their name
author_with(package_network, name = "Ioannis", exact = TRUE)
## Find all packages that package Rcpp suggests
suggests(package_network, package = "Rcpp", exact = TRUE)
## Find all packages that package Rcpp imports
imports(package_network, package = "Rcpp", exact = TRUE)
## Find all packages that package RcppArmadillo is linking to
linking_to(package_network, package = "RcppArmadillo", exact = TRUE)

## Using an author collaboration network
author_network <- build_network(cran_db, perspective = "author")
## Find all packages containing glm in their name
package_with(author_network, name = "glm")
## Find all authors of packages containing brglm in their name
author_of(author_network, package = "rglm", exact = FALSE)
## Find all packages with brglm in their name
package_with(author_network, name = "rglm", exact = FALSE)
## Find all authors of the package brglm2
author_of(author_network, package = "brglm2", exact = TRUE)
## Find all authors with Ioannis in their name
author_with(author_network, name = "Ioannis", exact = TRUE)
}
}
