% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_CRAN_db.R
\name{standardize_whitespace}
\alias{standardize_whitespace}
\title{Standardize whitespace in strings}
\usage{
standardize_whitespace(variable)
}
\arguments{
\item{variable}{a character string.}
}
\value{
A list of one vector of character strings.
}
\description{
Standardize whitespace in strings
}
\examples{
standardize_whitespace("  My spacebar         key is      broken.             ")

}
