% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cranly_network}
\alias{summary.cranly_network}
\alias{summary_cranly_network}
\title{Compute a range of author and package network statistics}
\usage{
\method{summary}{cranly_network}(object, ...)
}
\arguments{
\item{object}{a \code{\link{cranly_network}} object}

\item{...}{currently not used}
}
\description{
Compute a range of author and package network statistics
}
\details{

}
