% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.cranly_db_network}
\alias{visualize.cranly_db_network}
\title{Visualise a cranly_bd_network}
\usage{
visualize.cranly_db_network(object, packages = "MASS",
  physics_threshold = 300, height = "800px", width = "800px",
  types = c("import", "suggest", "enhance", "depend"))
}
\description{
Visualise a cranly_bd_network
}
\examples{

}
