% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\name{visualize}
\alias{visualize}
\alias{visualise}
\title{\code{visualize} method for an object}
\usage{
visualize(object, ...)

visualise(object, ...)
}
\arguments{
\item{object}{an object to be visualized}

\item{...}{other arguments to be passed to the method}
}
\description{
\code{visualize} method for an object
}
\seealso{
visualize.cranly_network
}
