% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cranly_network.R
\name{build_network.cranly_db}
\alias{build_network.cranly_db}
\alias{cranly_network}
\title{Extract edges and nodes from a \code{\link{cranly_db}} object}
\usage{
\method{build_network}{cranly_db}(object = clean_CRAN_db(), trace = FALSE,
  perspective = "package", ...)
}
\description{
Extract edges and nodes from a \code{\link{cranly_db}} object
}
\examples{
\dontrun{
package_db <- clean_CRAN_db()
package_network <- build_network(package_db)
head(package_network$edges)
head(package_network$nodes)
attr(package_network, "timestamp")
class(package_network)
}

}
