% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build_network.cranly_db}
\alias{build_network.cranly_db}
\alias{cranly_network}
\title{Compute edges and nodes of package and author networks}
\usage{
\method{build_network}{cranly_db}(object = clean_CRAN_db(), trace = FALSE,
  perspective = "package", ...)
}
\arguments{
\item{object}{a \code{\link{cranly_db}} object}

\item{trace}{logical. Print progress infromation? Default is \code{FALSE}}

\item{perspective}{character. Build a \code{"package"} (default )or an \code{"author"} network?}

\item{...}{Currently not used}
}
\description{
Compute edges and nodes of package and author networks
}
\details{
The convention for a \code{\link{cranly_network}} with
\code{perspective = "package"} is that the direction of an edge is
from the package that is imported by, suggested by, enhances or is
a dependency of another package, to the latter package. The author
network is analysed and visualised as undirected by all methods in
\code{cranly}.
}
\examples{
\dontrun{
data("cran20032018", package = "cranly")
## Package network
package_network <- build_network(object = cran20032018, perspective = "package")
head(package_network$edges)
head(package_network$nodes)
attr(package_network, "timestamp")
class(package_network)

## Author network
author_network <- build_network(object = cran20032018, perspective = "author")
head(author_network$edges)
head(author_network$nodes)
attr(author_network, "timestamp")
class(author_network)
}

}
