% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_CRAN_db.R
\name{clean_CRAN_db}
\alias{clean_CRAN_db}
\alias{cranly_db}
\title{Clean and organise directives and authors in the output of \code{tools::CRAN_package_db()}}
\usage{
clean_CRAN_db(packages_db = tools::CRAN_package_db(),
  clean_directives = clean_up_directives, clean_author = clean_up_author)
}
\description{
Clean and organise directives and authors in the output of \code{tools::CRAN_package_db()}
}
\examples{
\dontrun{
data("cran_db_example", package = "cranly")
package_db <- clean_CRAN_db(cran_db_example)
}
}
