% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranly-package.R
\docType{package}
\name{cranly}
\alias{cranly}
\alias{cranly-package}
\title{cranly: CRAN package database analytics and visualizations}
\description{
cranly: CRAN package database analytics and visualizations
}
\details{
\pkg{cranly} provides core visualizations and summaries for the
CRAN package database. The package provides comprehensive methods
for cleaning up and organizing the information in the CRAN package
database, for building package directives networks (depends,
imports, suggests, enhances, linking to) and collaboration
networks, and for computing summaries and producing interactive
visualizations from the resulting networks. Network visualization
is through the \pkg{visNetwork}
(\url{https://CRAN.R-project.org/package=visNetwork}) package. The
package also provides functions to coerce the networks to
\pkg{igraph} \url{https://CRAN.R-project.org/package=igraph}
objects for further analyses and modelling.
}
\section{Acknowledgements}{

\itemize{

\item David Selby (\url{http://selbydavid.com}) experimented with
and provided helpful comments and feedback on a pre-release version
of \pkg{cranly}. His help is gratefully acknowledged.

\item This work has been partially supported by the Alan Turing
Institute under the EPSRC grant EP/N510129/1 (Turing award number
TU/B/000082).

}
}

