% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary_cranly_network}
\alias{plot.summary_cranly_network}
\title{Top-n packge or author barplots according to a range of network statistics}
\usage{
\method{plot}{summary_cranly_network}(x, top = 10, according_to = NULL,
  scale = FALSE, ...)
}
\description{
Top-n packge or author barplots according to a range of network statistics
}
\examples{

\dontrun{
data("cran20032018", package = "cranly")
package_network <- build_network(object = cran20032018)
package_summaries <- summary(package_network)
plot(package_summaries, top = 30)
}

}
