% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cranly_network.R
\name{setup_cranly_network}
\alias{setup_cranly_network}
\title{Extract edges and nodes from a \code{\link{cranly_db}} object}
\usage{
setup_cranly_network(object = clean_CRAN_db(), trace = FALSE)
}
\description{
Extract edges and nodes from a \code{\link{cranly_db}} object
}
\examples{
\dontrun{
data("cran_db_example", package = "cranly")
package_db <- clean_CRAN_db(cran_db_example)
package_network <- setup_cranly_network(packages_db)
head(package_network$edges)
head(package_network$nodes)
attr(package_network, "timestamp")
class(package_network)
}

 @export
}
