% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.cranly_network}
\alias{subset.cranly_network}
\title{Subset a \code{\link{cranly_network}} according to author, package and/or directive}
\usage{
\method{subset}{cranly_network}(x, package = NULL, author = NULL,
  directive = c("imports", "suggests", "enhances", "depends"), exact = TRUE,
  only = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{cranly_network}} object}

\item{package}{a vector of character strings with the package names to be matched}

\item{author}{a vector of character strings with the author names to be matched}

\item{directive}{a vector of at least one of \code{"Imports"}, \code{"Suggests"}, \code{"Enhances"}, \code{"Depends"}}

\item{exact}{logical. Should we use exact matching? Default is \code{TRUE}}

\item{only}{logical. If \code{TRUE} the subset includes only the edges between packages named in \code{package} and/or authors named in \code{author}. If \code{FALSE} (default) edges to and from all other packages and/or authors that are linked to \code{package} and/or \code{author} are included in the subset}

\item{...}{currently not used}
}
\description{
Subset a \code{\link{cranly_network}} according to author, package and/or directive
}
