% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.cranly_network}
\alias{subset.cranly_network}
\title{Subset a \code{\link{cranly_network}} according to author, package and/or directive}
\usage{
\method{subset}{cranly_network}(x, package = "cranly",
  author = "Ioannis Kosmidis", directive = c("Imports", "Suggests",
  "Enhances", "Depends"), exact = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{cranly_network}} object}

\item{package}{a vector of character strings with the package names to be matched}

\item{author}{a vector of character strings with the author names to be matched}

\item{directive}{a vector of at least one of \code{"Imports"}, \code{"Suggests"}, \code{"Enhances"}, \code{"Depends"}}

\item{exact}{logical. Should we use exact matching of author and package names? Default is \code{TRUE}}

\item{...}{currently not used}
}
\description{
Subset a \code{\link{cranly_network}} according to author, package and/or directive
}
