% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.cranly_network}
\alias{visualize.cranly_network}
\title{Visualise a \code{\link{cranly_network}}}
\usage{
\method{visualize}{cranly_network}(object, package = NULL, author = NULL,
  physics_threshold = 200, height = NULL, width = NULL,
  edge_type = c("Imports", "Suggests", "Enhances", "Depends"),
  dragNodes = TRUE, dragView = TRUE, zoomView = TRUE, ...)
}
\description{
Visualise a \code{\link{cranly_network}}
}
